+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 0.8.0
+### Changed
+- Peers are now killed directly instead of through peer manager.
+
+### Removed
+- Chain no longer needs peer manager.
+
 ## 0.7.2
 ### Added
 - Compatibility with base 4.12.